replace int with double of t_points::x,y,z

+++ inc/fdf.h
@@ -6,7 +6,7 @@
 /*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/23 19:59:13 by reasuke           #+#    #+#             */
-/*   Updated: 2023/12/15 19:19:23 by reasuke          ###   ########.fr       */
+/*   Updated: 2023/12/18 17:52:12 by reasuke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -51,10 +51,10 @@ typedef struct s_mlx
 
 typedef struct s_point
 {
-	int	x;
-	int	y;
-	int	z;
-	int	color;
+	double	x;
+	double	y;
+	double	z;
+	int		color;
 }	t_point;
 
 typedef struct s_map
@@ -96,7 +96,7 @@ void	render_image(t_map *map, t_mlx *mlx);
 
 void	exit_with_error(char *message);
 
-void	translate_points(t_map *map, int dx, int dy, int dz);
+void	translate_points(t_map *map, double dx, double dy, double dz);
 void	scale_points(t_map *map, double scale_factor, int axis_flag);
 void	rotate_points_x(t_map *map, double theta);
 void	rotate_points_y(t_map *map, double theta);
